pyproject.toml
[project]
name = "recall-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.184.0",
    "google-auth>=2.41.1",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "livekit>=1.0.13",
    "livekit-agents>=1.2.14",
    "livekit-plugins-cartesia>=1.2.14",
    "livekit-plugins-deepgram>=1.2.14",
    "livekit-plugins-openai>=1.2.14",
    "livekit-plugins-silero>=1.2.14",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
]
